Move utils, so it's installable.
